+++ SuperNewRoles/CustomCosmetics/CustomCosmeticsMenus/Patch/ObjectData.cs
@@ -1,6 +1,4 @@
-using System;
 using System.Collections.Generic;
-using System.Text;
 using BepInEx.Configuration;
 using TMPro;
 using UnityEngine;
