+++ SuperNewRoles/CustomCosmetics/CustomCosmeticsMenus/Patch/SelectPatch.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
 using HarmonyLib;
 using static SuperNewRoles.CustomCosmetics.CustomCosmeticsMenus.Patch.ObjectData;
