+++ SuperNewRoles/CustomCosmetics/CustomCosmeticsMenus/Patch/UpdatePatch.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Text;
 using HarmonyLib;
 using UnityEngine;
