--- SuperNewRoles/CustomCosmetics/ShareCosmetics/DownLoadVisor.cs
@@ -1,7 +1,0 @@
-
-namespace SuperNewRoles.CustomCosmetics.ShareCosmetics
-{
-    class DownLoadVisor
-    {
-    }
-}
