--- SuperNewRoles/CustomCosmetics/ShareCosmetics/ShareNamePlate.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace SuperNewRoles.CustomCosmetics.ShareCosmetics
-{
-    class ShareNamePlate
-    {
-        public static Dictionary<int, Sprite> NamePlateSprites;
-    }
-}
