--- SuperNewRoles/EndGame/CustomEndReason.cs
@@ -1,27 +1,0 @@
-
-
-namespace SuperNewRoles.EndGame
-{
-    public enum CustomGameOverReason
-    {
-        CrewmateWin = 7,
-        ImpostorWin = 8,
-        GodWin = 9,
-        HAISON = 10,
-        JesterWin,
-        JackalWin,
-        QuarreledWin,
-        EgoistWin,
-        WorkpersonWin,
-        LoversWin,
-        MadJesterWin,
-        FalseChargesWin,
-        FoxWin,
-        DemonWin,
-        ArsonistWin,
-        VultureWin,
-        TunaWin,
-        NeetWin,
-        BugEnd
-    }
-}
