+++ SuperNewRoles/EndGame/EndGame.cs
@@ -15,6 +15,28 @@ using UnityEngine;
 
 namespace SuperNewRoles.EndGame
 {
+    public enum CustomGameOverReason
+    {
+        CrewmateWin = 7,
+        ImpostorWin = 8,
+        GodWin = 9,
+        HAISON = 10,
+        JesterWin,
+        JackalWin,
+        QuarreledWin,
+        EgoistWin,
+        WorkpersonWin,
+        LoversWin,
+        MadJesterWin,
+        FalseChargesWin,
+        FoxWin,
+        DemonWin,
+        ArsonistWin,
+        VultureWin,
+        TunaWin,
+        NeetWin,
+        BugEnd
+    }
     enum WinCondition
     {
         Default,
